+++ es/processors/loaders/index.js
@@ -1,7 +1,8 @@
 const StdinReader = require('./StdinReader');
 const HttpClient = require('./HttpClient');
 const FileReader = require('./FileReader');
-const CsvLoader = require('./CsvFileLoader');
+const CsvFileLoader = require('./CsvFileLoader');
+const Browser = require('./Browser');
 const IdentityProcessor = require('../IdentityProcessor');
 const NoOperationProcessor = require('../NoOperationProcessor');
 
@@ -14,7 +15,8 @@ module.exports = {
   stdin: StdinReader,
   http: HttpClient,
   file: FileReader,
-  'csv-file': CsvLoader,
+  'csv-file': CsvFileLoader,
+  browser: Browser,
   identity: IdentityProcessor,
   noop: NoOperationProcessor,
 };